General vimrc cleanup
Changes in this commit should be probably introduced in separate
commits, but fuck it. This commit represents big clean up of unused
options and remove all settings which were related misusing vim tabs in
a way for which buffers were designed. This is now over and probably
the tabs were not be used at all in further workflow.

Just one blog reference which served as a final push:

http://joshldavis.com/2014/04/05/vim-tab-madness-buffers-vs-tabs/